Refactor rust code for python bindings
